Product is in reasonable state. Slight changes to lighting and restrictions of space should allow for a demoable product.
Added: FlippedCameraImageMaterial and FlippedImageSetter to accommodate horizontally flppped image produced from NV21 conversion.
Fix: Setting the Texture displayed on the image plane to Texture2D.black caused a strange crash - removed the texture setting.
Added: Behavior: Make plane and cloaking box invisible when the user does not have a good pose position to avoid drastic skewing of the image.
Fix: PlaneRect was the root of the strange image distortion upon rotation that was occurring. Same fix applied as in POV_Portal project where the construction of the rectangular plane using min-max points were corrected.
Added: Some flexibility in the PoseVoxelSet editor script to make it so that the dimensions were easily visible and adjustable in the Editor. This makes it less tedious to hard code the values while I search for a reasonable compromise between availability and accuracy.
Added: Camera Manager script tweaks to make it easier for me to adjust culling masks on the fly in the Editor instead of having to hard code the values. Helped with debugging plane issues.
Fix: Slight scrunching and displacement of image derived from image plane was due to the Tango camera prefab having the wrong metrics compared to the hardware. This meant that a prebaked plane could come out too short in comparison to the width of the viewport, causing images taken from the viewport to get scrunched up.